MM-31968 Save auto response props on keyboard done
