+++ app/screens/settings/notification_settings_mentions/notification_settings_mention_base.js
@@ -111,8 +111,6 @@ export default class NotificationSettingsMentionsBase extends PureComponent {
         }
 
         mentionKeys = mentionKeys.join(',');
-        Reflect.deleteProperty(notifyProps, 'showKeywordsModal');
-        Reflect.deleteProperty(notifyProps, 'showReplyModal');
 
         this.props.onBack({
             ...notifyProps,
